renaming type from requirements to pypi, to match the existing projects.json
