+++ src/lib/dependencies/dependency-managers.js
@@ -3,7 +3,7 @@ import * as npm from './npm';
 import * as nuget from './nuget';
 import * as dep from './dep';
 import * as bundler from './bundler';
-import * as requirements from './requirements';
+import * as pypi from './pypi';
 
 const dependencyManagers = {
   npm,
@@ -11,7 +11,7 @@ const dependencyManagers = {
   nuget,
   dep,
   bundler,
-  requirements,
+  pypi,
 };
 
 export default dependencyManagers;
